add: struture and camelCase
