+++ common/validate.go
@@ -3,14 +3,13 @@ package common
 import "github.com/klassmann/cpfcnpj"
 
 func ValidatingCPF(cpf string) bool {
-	cpf_clear := cpfcnpj.Clean(cpf)
-	isValid := cpfcnpj.ValidateCPF(cpf_clear)
+	cpfClear := cpfcnpj.Clean(cpf)
+	isValid := cpfcnpj.ValidateCPF(cpfClear)
 	return isValid
 }
 
 func ValidatingCNPJ(cnpj string) bool {
-	cnpj_clear := cpfcnpj.Clean(cnpj)
-	isValid := cpfcnpj.ValidateCNPJ(cnpj_clear)
+	cnpjClear := cpfcnpj.Clean(cnpj)
+	isValid := cpfcnpj.ValidateCNPJ(cnpjClear)
 	return isValid
 }
-
